add brands and optimiza code

+++ controllers/category.controller.js
@@ -14,6 +14,12 @@ exports.getAllCategories = async (req, res) => {
 };
 exports.createCategory = async (req, res) => {
 	try {
+		if (req.userId.user_level === 0) {
+			return res.status(400).json({
+				status: "error",
+				message: "You are not authorized to perform this action",
+			});
+		}
 		const category = new Category(req.body);
 		await category.save();
 		return res.status(200).json({
@@ -27,6 +33,12 @@ exports.createCategory = async (req, res) => {
 };
 exports.deleteCategory = async (req, res) => {
 	try {
+		if (req.userId.user_level === 0) {
+			return res.status(400).json({
+				status: "error",
+				message: "You are not authorized to perform this action",
+			});
+		}
 		const category = await Category.findByIdAndDelete(req.params.id);
 		return res.status(200).json({
 			status: "success",
@@ -39,17 +51,21 @@ exports.deleteCategory = async (req, res) => {
 };
 exports.updateCategory = async (req, res) => {
 	try {
-    const category = await Category.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      { new: true }
-    );
-    return res.status(200).json({
-      status: "success",
-      message: "Category updated successfully",
-      data: category,
-    });
-  } catch (error) {
-    return res.status(500).json({message: error.message});
-  }
-}
+		if (req.userId.user_level === 0) {
+			return res.status(400).json({
+				status: "error",
+				message: "You are not authorized to perform this action",
+			});
+		}
+		const category = await Category.findByIdAndUpdate(req.params.id, req.body, {
+			new: true,
+		});
+		return res.status(200).json({
+			status: "success",
+			message: "Category updated successfully",
+			data: category,
+		});
+	} catch (error) {
+		return res.status(500).json({message: error.message});
+	}
+};
